feat: use factored out ratio-proof/pairing
