+++ manta-trusted-setup/Cargo.toml
@@ -35,7 +35,7 @@ std = []
 ark-ec = { version = "0.3.0", default-features = false }
 ark-ff = { version = "0.3.0", default-features = false }
 ark-groth16 = { version = "0.3.0", default-features = false }
-ark-poly = {version = "0.3.0", default-features = false}
+ark-poly = { version = "0.3.0", default-features = false }
 ark-relations = { version = "0.3.0", default-features = false }
 ark-serialize = { version = "0.3.0", default-features = false, features = ["derive"] }
 ark-snark = { version = "0.3.0", default-features = false }
@@ -51,6 +51,6 @@ rayon = { version = "1.5.3", optional = true, default-features = false }
 workspace-hack = { version = "0.1.0", path = "../workspace-hack" }
 
 [dev-dependencies]
-ark-bls12-381 = {version = "0.3.0", default-features = false, features = ["curve", "scalar_field"]}
+ark-bls12-381 = { version = "0.3.0", default-features = false, features = ["curve", "scalar_field"] }
 ark-r1cs-std = { version = "0.3.1", default-features = false }
 manta-pay = { path = "../manta-pay", default-features = false, features = ["groth16"] } # TODO: To be removed
