+++ manta-trusted-setup/src/lib.rs
@@ -25,4 +25,6 @@ extern crate alloc;
 
 pub mod groth16;
 pub mod mpc;
+pub mod pairing;
+pub mod ratio;
 pub mod util;
